sort out workflow, to match upcoming v4
